CORE-12562 Add internal functionality to resolv stateref into UtxoTransactionOutputDto and let us deseriaize them to stateandref
