+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/persistence/external/events/FindUnconsumedStatesByTypeExternalEventFactory.kt
@@ -8,7 +8,7 @@ import net.corda.data.ledger.persistence.UtxoTransactionOutputs
 import net.corda.flow.external.events.factory.ExternalEventFactory
 import net.corda.flow.external.events.factory.ExternalEventRecord
 import net.corda.flow.state.FlowCheckpoint
-import net.corda.ledger.utxo.flow.impl.persistence.UtxoTransactionOutputDto
+import net.corda.ledger.utxo.data.transaction.UtxoTransactionOutputDto
 import net.corda.schema.Schemas
 import net.corda.virtualnode.toAvro
 import org.osgi.service.component.annotations.Activate
